combined the enemy base with the path system
