--- Assets/Max Test/Circle.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 6bdd4e07ad2520e4c95d0f89ca58714b
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
