--- Assets/Max Test/LineFollower.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 02b422d1747e6fa4284a1dfbfc2f17a7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
