+++ Assets/Max Test/w1b2.asset
@@ -12,7 +12,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: a5ed32b1dd0952740acff367c060210d, type: 3}
   m_Name: w1b2
   m_EditorClassIdentifier: 
-  enemy: {fileID: 7660246423821641570, guid: 5eaf02e2da541e247a28c49ba928a329, type: 3}
+  enemy: {fileID: 7177020029656068024, guid: 609511b6d7251f04985d0a52fa4645d4, type: 3}
   numberOfEnemy: 15
   enemyDelay: 1
   batchDelay: 0
