--- Assets/Max Test/y.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 5eaf02e2da541e247a28c49ba928a329
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
